проверка прав на роль employee

+++ src/main/java/com/OnlineBankingService/service/impl/CreditTariffServiceImpl.java
@@ -1,8 +1,11 @@
 package com.OnlineBankingService.service.impl;
 
+import com.OnlineBankingService.config.RestTemplateConfig;
 import com.OnlineBankingService.entity.CreditTariff;
+import com.OnlineBankingService.entity.dto.AuthValidationRequest;
 import com.OnlineBankingService.entity.dto.CreditTariffRequest;
 import com.OnlineBankingService.entity.dto.CreditTariffResponse;
+import com.OnlineBankingService.entity.dto.DeleteCreditTariffRequest;
 import com.OnlineBankingService.repository.CreditTariffRepository;
 import com.OnlineBankingService.service.CreditTariffService;
 import lombok.RequiredArgsConstructor;
@@ -11,7 +14,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.domain.Specification;
-import org.springframework.http.HttpStatus;
+import org.springframework.http.*;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -26,9 +29,26 @@ import java.util.UUID;
 public class CreditTariffServiceImpl implements CreditTariffService {
 
     private final CreditTariffRepository creditTariffRepository;
+    private final RestTemplateConfig restTemplateConfig;
 
     @Override
     public CreditTariffResponse createCreditTariff (CreditTariffRequest request) {
+
+        AuthValidationRequest authValidationRequest = AuthValidationRequest.builder()
+                .token(request.getToken())
+                .build();
+
+        HttpHeaders headers = new HttpHeaders();
+        headers.setContentType(MediaType.APPLICATION_JSON);
+
+        HttpEntity<AuthValidationRequest> entity = new HttpEntity<>(authValidationRequest, headers);
+
+        ResponseEntity<Boolean> response = restTemplateConfig.restTemplate().postForEntity("http://localhost:8085/api/auth/validate-employee", entity, Boolean.class);
+
+        if (!response.getStatusCode().is2xxSuccessful() || Boolean.FALSE.equals(response.getBody())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Ошибка при проверке сотрудника");
+        }
+
         CreditTariff creditTariff = CreditTariff.builder()
                 .name(request.getName())
                 .description(request.getDescription())
@@ -50,7 +70,22 @@ public class CreditTariffServiceImpl implements CreditTariffService {
     }
 
     @Override
-    public void deleteCreditTariff (UUID id){
+    public void deleteCreditTariff (UUID id, DeleteCreditTariffRequest request){
+        AuthValidationRequest authValidationRequest = AuthValidationRequest.builder()
+                .token(request.getToken())
+                .build();
+
+        HttpHeaders headers = new HttpHeaders();
+        headers.setContentType(MediaType.APPLICATION_JSON);
+
+        HttpEntity<AuthValidationRequest> entity = new HttpEntity<>(authValidationRequest, headers);
+
+        ResponseEntity<Boolean> response = restTemplateConfig.restTemplate().postForEntity("http://localhost:8085/api/auth/validate-employee", entity, Boolean.class);
+
+        if (!response.getStatusCode().is2xxSuccessful() || Boolean.FALSE.equals(response.getBody())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Ошибка при проверке сотрудника");
+        }
+
         creditTariffRepository.deleteById(id);
     }
 
@@ -133,6 +168,21 @@ public class CreditTariffServiceImpl implements CreditTariffService {
 
     @Override
     public CreditTariffResponse updateCreditTariff(UUID id, CreditTariffRequest request){
+
+        AuthValidationRequest authValidationRequest = AuthValidationRequest.builder()
+                .token(request.getToken())
+                .build();
+
+        HttpHeaders headers = new HttpHeaders();
+        headers.setContentType(MediaType.APPLICATION_JSON);
+
+        HttpEntity<AuthValidationRequest> entity = new HttpEntity<>(authValidationRequest, headers);
+
+        ResponseEntity<Boolean> response = restTemplateConfig.restTemplate().postForEntity("http://localhost:8085/api/auth/validate-employee", entity, Boolean.class);
+
+        if (!response.getStatusCode().is2xxSuccessful() || Boolean.FALSE.equals(response.getBody())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Ошибка при проверке сотрудника");
+        }
 
         CreditTariff creditTariff = creditTariffRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,"Данного Кредитного тарифа не существует"));
